+++ ChangeLog.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Not Released
+- NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *11/05/2022* 2.5.0
 
 ## Version 2.4